adding rtl8169 driver I guess we doing networking now (even tho I didn't finish the ahci one yet but meh)
this (should) work for receiving packets, still gotta test it and implement sending packets

+++ src/pci/pci.h
@@ -49,7 +49,7 @@ typedef struct pci_dev {
     // other extracted info
     pci_bar_t* bars;
     pci_cap_t* caps;
-    uint8_t irq;
+    int irq;
 
     // the mmio base address
     char* mmio;
